Move common writing code in a doWrite method in the parent class
